# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Interview Reader API
  version: 0.2.0
  description: Reader study APIs and isolated desktop administration APIs.
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /admin/documents:
    get:
      summary: List documents and draft counts for desktop management
      parameters:
        - { name: query, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminDocumentPage" }
  /admin/documents/{documentId}:
    get:
      summary: Get one document with version and draft counts
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminDocumentSummary" }
        "404": { $ref: "#/components/responses/Problem" }
  /admin/documents/{documentId}/metadata:
    get:
      summary: Get independently editable document metadata
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentMetadata" }
    patch:
      summary: Update document metadata with optimistic locking; the document code remains read-only
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDocumentMetadataRequest" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Metadata updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentMetadata" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/documents/{documentId}/versions:
    get:
      summary: List published, retired, and draft versions
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VersionSummary" }
  /admin/documents/{documentId}/take-down:
    post:
      summary: Hide a published document from the reader while preserving all data
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "204": { description: Taken down }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/documents/{documentId}/restore:
    post:
      summary: Restore an offline document using its retained published version
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "204": { description: Restored }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/documents/{documentId}/deletion:
    post:
      summary: Start an irreversible permanent document deletion
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteDocumentRequest" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "202":
          description: Deletion accepted or existing idempotent job returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletionJob" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/document-deletions/{jobId}:
    get:
      summary: Get durable permanent deletion progress
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Deletion progress
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletionJob" }
  /admin/document-deletions/{jobId}/retry:
    post:
      summary: Retry a failed permanent deletion without repeating title confirmation
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "202":
          description: Retry accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletionJob" }
  /admin/documents/{documentId}/versions/{sourceVersionId}/revisions:
    post:
      summary: Create an editable draft from a selected version
      parameters:
        - $ref: "#/components/parameters/DocumentId"
        - { name: sourceVersionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "201":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VersionSummary" }
  /admin/versions/{versionId}/editor:
    get:
      summary: Load draft metadata and the node tree without block bodies
      parameters:
        - $ref: "#/components/parameters/VersionId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EditorSnapshot" }
    delete:
      summary: Delete an unpublished draft
      parameters:
        - $ref: "#/components/parameters/VersionId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "204": { description: Deleted }
  /admin/versions/{versionId}/editor/nodes/{nodeId}/blocks:
    get:
      summary: Load one page of blocks for a draft node
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - $ref: "#/components/parameters/NodeId"
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 40 } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeBlocksPage" }
    post:
      summary: Append a block to a draft node
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - $ref: "#/components/parameters/NodeId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlockWriteRequest" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "201":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EditorBlock" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/versions/{versionId}/editor/nodes/{nodeId}:
    patch:
      summary: Update draft node properties
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - $ref: "#/components/parameters/NodeId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateNodeRequest" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EditorSnapshot" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      summary: Delete an empty draft node and promote its direct children
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - $ref: "#/components/parameters/NodeId"
        - { name: draftRevision, in: query, required: true, schema: { type: integer, minimum: 0 } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Updated editor snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EditorSnapshot" }
        "409":
          description: Draft revision conflict, non-empty node, or last remaining node
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /admin/versions/{versionId}/editor/structure:
    patch:
      summary: Replace parent and sibling order for all draft nodes
      parameters:
        - $ref: "#/components/parameters/VersionId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StructureUpdateRequest" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EditorSnapshot" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/versions/{versionId}/editor/blocks/{blockId}:
    patch:
      summary: Update one draft block
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - { name: blockId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlockWriteRequest" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EditorBlock" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      summary: Delete one draft block
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - { name: blockId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: draftRevision, in: query, required: true, schema: { type: integer, minimum: 0 } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockMutationResult" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/versions/{versionId}/editor/blocks/{blockId}/image:
    post:
      summary: Upload a verified image and atomically attach it to a draft image block
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - { name: blockId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, draftRevision]
              properties:
                file: { type: string, format: binary }
                draftRevision: { type: integer, minimum: 0 }
                alt: { type: string, maxLength: 300 }
                decorative: { type: boolean, default: false }
                caption: { type: string, maxLength: 1000 }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Image attached to the block
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImageBlockUploadResult" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/versions/{versionId}/editor/assets/{assetKey}:
    get:
      summary: Read one private image from a draft for editor preview
      parameters:
        - $ref: "#/components/parameters/VersionId"
        - { name: assetKey, in: path, required: true, schema: { type: string } }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200": { description: Image bytes }
        "404": { $ref: "#/components/responses/Problem" }
  /admin/versions/{versionId}/editor/blocks/cleanup-empty:
    post:
      summary: Delete empty blocks and resequence affected nodes
      parameters:
        - $ref: "#/components/parameters/VersionId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlockCleanupRequest" }
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "200":
          description: Successful response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockMutationResult" }
        "409": { $ref: "#/components/responses/Problem" }
  /admin/documents/{documentId}/versions/{versionId}/publish:
    post:
      summary: Publish a draft and retire the previous published version
      parameters:
        - $ref: "#/components/parameters/DocumentId"
        - $ref: "#/components/parameters/VersionId"
      responses:
        default: { $ref: "#/components/responses/Problem" }
        "204": { description: Published }
        "409": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: IR_SESSION
  parameters:
    DocumentId: { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
    VersionId: { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
    NodeId: { name: nodeId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Problem:
      description: RFC 9457-compatible problem detail
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
    Uuid:
      type: string
      format: uuid
    NullableUuid:
      type: [string, "null"]
      format: uuid
    SourceType:
      type: string
      enum: [PDF, EXCEL, JSON_PACKAGE, MARKDOWN, MANUAL]
    NodeType:
      type: string
      enum: [PART, CHAPTER, SECTION, SUBSECTION, QUESTION, APPENDIX, OTHER]
    BlockType:
      type: string
      enum: [paragraph, heading_note, unordered_list, ordered_list, code, table, quote, callout, formula, image, divider, table_snapshot]
    SemanticRole:
      type: string
      enum: [QUESTION, ANSWER, EXPLANATION, CONCLUSION, INTRODUCTION, DIRECTORY, PRINCIPLE, PRACTICE, PITFALL, FOLLOW_UP]
    VersionStatus:
      type: string
      enum: [DRAFT, PUBLISHED, RETIRED]
    DocumentStatus:
      type: string
      enum: [DRAFT, PUBLISHED, OFFLINE, DELETING, DELETE_FAILED]
    DeletionJobStatus:
      type: string
      enum: [QUEUED, RUNNING, FAILED, COMPLETED]
    DeletionStage:
      type: string
      enum: [QUEUED, CLIENT_SYNC_MARKED, DELETING_FILES, DELETING_DATA, COMPLETED, FAILED]
...
    Problem:
      type: object
      required: [type, title, status, detail, code, error, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string }
        error: { type: string }
        traceId: { type: string, format: uuid }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
...
    SourceBbox:
      type: object
      additionalProperties: true
      properties:
        page: { type: [integer, "null"] }
        x: { type: [number, "null"] }
        y: { type: [number, "null"] }
        width: { type: [number, "null"] }
        height: { type: [number, "null"] }
...
    DeleteDocumentRequest:
      type: object
      required: [confirmationTitle]
      properties:
        confirmationTitle: { type: string }
    DeletionJob:
      type: object
      required: [id, documentId, status, currentStage, attemptCount, errorCode, errorMessage, requestedAt, startedAt, completedAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        documentId: { $ref: "#/components/schemas/Uuid" }
        status: { $ref: "#/components/schemas/DeletionJobStatus" }
        currentStage: { $ref: "#/components/schemas/DeletionStage" }
        attemptCount: { type: integer, minimum: 0 }
        errorCode: { type: [string, "null"] }
        errorMessage: { type: [string, "null"] }
        requestedAt: { type: string, format: date-time }
        startedAt: { type: [string, "null"], format: date-time }
        completedAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
...
    DocumentMetadata:
      type: object
      required: [documentId, code, title, description, tags, metadataRevision, duplicateTitleCount]
      properties:
        documentId: { $ref: "#/components/schemas/Uuid" }
        code: { type: string, readOnly: true }
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, "null"], maxLength: 5000 }
        tags:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 50 }
        metadataRevision: { type: integer, minimum: 0 }
        duplicateTitleCount: { type: integer, minimum: 0 }
    UpdateDocumentMetadataRequest:
      type: object
      required: [metadataRevision, title, description, tags]
      properties:
        metadataRevision: { type: integer, minimum: 0 }
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, "null"], maxLength: 5000 }
        tags:
          type: array
          maxItems: 20
          items: { type: string, minLength: 1, maxLength: 50 }
    AdminDocumentSummary:
      type: object
      required: [id, code, title, status, currentVersionId, versionCount, draftCount, updatedAt, deletionJob]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        code: { type: string }
        title: { type: string }
        status: { $ref: "#/components/schemas/DocumentStatus" }
        currentVersionId: { $ref: "#/components/schemas/NullableUuid" }
        versionCount: { type: integer, minimum: 0 }
        draftCount: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
        deletionJob:
          anyOf: [{ $ref: "#/components/schemas/DeletionJob" }, { type: "null" }]
    AdminDocumentPage:
      type: object
      required: [items, page, size, hasNext]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AdminDocumentSummary" } }
        page: { type: integer, minimum: 1 }
        size: { type: integer, minimum: 1, maximum: 100 }
        hasNext: { type: boolean }
    VersionSummary:
      type: object
      required: [id, versionNo, parentVersionId, parentVersionNo, originImportJobId, sourceType, sourceFileName, status, draftRevision, publishedAt, createdAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        versionNo: { type: integer, minimum: 1 }
        parentVersionId: { $ref: "#/components/schemas/NullableUuid" }
        parentVersionNo: { type: [integer, "null"], minimum: 1 }
        originImportJobId: { $ref: "#/components/schemas/NullableUuid" }
        sourceType: { $ref: "#/components/schemas/SourceType" }
        sourceFileName: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/VersionStatus" }
        draftRevision: { type: integer, minimum: 0 }
        publishedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    EditorDocument:
      type: object
      required: [id, code, title, description, language]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        code: { type: string }
        title: { type: string }
        description: { type: [string, "null"] }
        language: { type: string }
    EditorNode:
      type: object
      required: [id, parentId, nodeKey, nodeType, semanticRole, title, level, sortOrder, anchor, sourcePageStart, sourcePageEnd]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        parentId: { $ref: "#/components/schemas/NullableUuid" }
        nodeKey: { type: string }
        nodeType: { $ref: "#/components/schemas/NodeType" }
        semanticRole:
          anyOf: [{ $ref: "#/components/schemas/SemanticRole" }, { type: "null" }]
        title: { type: string }
        level: { type: integer, minimum: 1 }
        sortOrder: { type: integer }
        anchor: { type: string }
        sourcePageStart: { type: [integer, "null"] }
        sourcePageEnd: { type: [integer, "null"] }
    EditorSnapshot:
      type: object
      required: [version, document, nodes]
      properties:
        version: { $ref: "#/components/schemas/VersionSummary" }
        document: { $ref: "#/components/schemas/EditorDocument" }
        nodes: { type: array, items: { $ref: "#/components/schemas/EditorNode" } }
    EditorBlock:
      type: object
      required: [id, blockKey, seq, blockType, payload, plainText, language, sourcePage, sourceBbox, confidence]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        blockKey: { type: string }
        seq: { type: integer }
        blockType: { $ref: "#/components/schemas/BlockType" }
        payload: { $ref: "#/components/schemas/JsonObject" }
        plainText: { type: string }
        language: { type: [string, "null"] }
        sourcePage: { type: [integer, "null"] }
        sourceBbox:
          anyOf: [{ $ref: "#/components/schemas/SourceBbox" }, { type: "null" }]
        confidence: { type: [number, "null"], minimum: 0, maximum: 1 }
    NodeBlocksPage:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/EditorBlock" } }
        nextCursor: { type: [string, "null"] }
    BlockMutationResult:
      type: object
      required: [draftRevision, removedCount]
      properties:
        draftRevision: { type: integer, minimum: 0 }
        removedCount: { type: integer, minimum: 0 }
    ImageBlockUploadResult:
      type: object
      required: [block, draftRevision]
      properties:
        block: { $ref: "#/components/schemas/EditorBlock" }
        draftRevision: { type: integer, minimum: 0 }
...
    UpdateNodeRequest:
      type: object
      required: [draftRevision, title, nodeType, semanticRole]
      properties:
        draftRevision: { type: integer, minimum: 0 }
        title: { type: string, minLength: 1 }
        nodeType: { $ref: "#/components/schemas/NodeType" }
        semanticRole:
          anyOf: [{ $ref: "#/components/schemas/SemanticRole" }, { type: "null" }]
    StructureNode:
      type: object
      required: [id, parentId, sortOrder]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        parentId: { $ref: "#/components/schemas/NullableUuid" }
        sortOrder: { type: integer }
    StructureUpdateRequest:
      type: object
      required: [draftRevision, nodes]
      properties:
        draftRevision: { type: integer, minimum: 0 }
        nodes: { type: array, minItems: 1, items: { $ref: "#/components/schemas/StructureNode" } }
    BlockWriteRequest:
      type: object
      required: [draftRevision, blockType, payload, plainText, language]
      properties:
        draftRevision: { type: integer, minimum: 0 }
        blockType: { $ref: "#/components/schemas/BlockType" }
        payload:
          anyOf: [{ $ref: "#/components/schemas/JsonObject" }, { type: "null" }]
        plainText: { type: [string, "null"] }
        language: { type: [string, "null"] }
    BlockCleanupRequest:
      type: object
      required: [draftRevision]
      properties:
        draftRevision: { type: integer, minimum: 0 }
...